// todo_app/climax/src/store/todo/types.ts
import { DeepReadonly } from 'vue';

export type Status = 'waiting' | 'working' | 'completed' | 'pending';

export interface Todo {
  id: number;
  title: string;
  status: Status;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface TodoState {
  todos: Todo[];
}

export type Params = Pick<Todo, 'title' | 'content' | 'status'>;

export interface TodoStore {
  state: DeepReadonly<TodoState>;
  getTodo: (id: number) => void;
  addTodo: (todo: Params) => void;
  updateTodo: (id: number, todo: Todo) => void;
  deleteTodo: (id: number) => void;
}